[unit] mp.o batch 13: menu draw/triangle/widescreen/cross/select/up virtuals, MultiplayerMgr game-state forwarders, session request/player-state handlers, item handle helpers, voice manager ctor + seqIDInList, frontend/game-listing statics (452/770)

+++ src/game/platform_xbox/XboxLiveMenus.cpp
@@ -675,7 +675,7 @@ LABEL_9:
     {
         if (!MPUIInterface::mLiveQueryActive || !MPUIInterface::mQueryFromID)
         {
-            unsigned int numGames = 0;
+              unsigned long numGames = 0;
             MPUIInterface::GameListingGet(numGames);
             if (numGames == 0)
             {
